30 bearer authentication using django rest framework

+++ ddpui/api/admin_controller.py
@@ -4,13 +4,14 @@ from ninja.errors import HttpError
 from ninja.errors import ValidationError
 from ninja.responses import Response
 from pydantic.error_wrappers import ValidationError as PydanticValidationError
+from rest_framework.authtoken import views
 
 from ddpui.utils.ddp_logger import logger
-from ddpui.auth import AdminAuthBearer, LoginData
+from ddpui.auth import AuthBearer
 from ddpui.models.org import Org, OrgSchema
 
 from ddpui.models.org_user import OrgUser, OrgUserResponse, OrgUserUpdate
-from ddpui.models.admin_user import AdminUser, AdminUserResponse
+from ddpui.models.admin_user import AdminUserResponse
 
 adminapi = NinjaAPI(urls_namespace="admin")
 
@@ -40,65 +41,68 @@ def ninja_default_error_handler(request, exc: Exception):
 
 
 @adminapi.post("/login/")
-def post_login(request, payload: LoginData):
-    """Docstring"""
-    if payload.password == "password":
-        user = AdminUser.objects.filter(email=payload.email).first()
-        if user:
-            token = f"fake-admin-auth-token:{user.id}"
-            logger.info("returning auth token " + token)
-            return {"token": token}
-    raise HttpError(400, "login denied")
+def post_login(request):
+    """Uses the username and password in the request to return an auth token"""
+    token = views.obtain_auth_token(request)
+    return token
 
 
-@adminapi.get("/getadminuser", response=AdminUserResponse, auth=AdminAuthBearer())
+@adminapi.get("/getadminuser", response=AdminUserResponse, auth=AuthBearer())
 def get_admin_user(request):
-    """Fetch the current admin logged in user"""
-    return request.auth
+    """return the admin user who made this request"""
+    return AdminUserResponse(
+        email=request.user.email,
+        active=request.user.is_active,
+    )
 
 
-@adminapi.get(
-    "/organizations/users", response=List[OrgUserResponse], auth=AdminAuthBearer()
-)
-def get_organization_users(request, org: str = None):
+@adminapi.get("/organizations/users", response=List[OrgUserResponse], auth=AuthBearer())
+def get_organization_users(request, orgname: str = None):
     """Fetch all organization users"""
     assert request.auth
-    query = OrgUser.objects.filter(active=True)
-    if org:
-        query = query.filter(org__name=org)
-    return query
+    if request.adminuser is None:
+        raise HttpError(400, "unauthorized")
+    query = OrgUser.objects.filter(user__is_active=True)
+    if orgname:
+        query = query.filter(org__name=orgname)
+    return [
+        OrgUserResponse(email=orguser.user.email, active=orguser.user.is_active)
+        for orguser in query
+    ]
 
 
 @adminapi.put(
-    "/organizations/users/{orguserid}", response=OrgUserResponse, auth=AdminAuthBearer()
+    "/organizations/users/{orguserid}", response=OrgUserResponse, auth=AuthBearer()
 )
 def put_organization_user(request, orguserid: int, payload: OrgUserUpdate):
-    """Edit an organization user"""
+    """update attributes of an orguser (or of the linked django user)"""
     assert request.auth
-    user = OrgUser.objects.filter(id=orguserid).first()
-    if user is None:
-        raise HttpError(400, "no such user id")
+    if request.adminuser is None:
+        raise HttpError(400, "unauthorized")
+    orguser = OrgUser.objects.filter(id=orguserid).first()
+    if orguser is None:
+        raise HttpError(400, "no such orguser id")
     if payload.email:
-        user.email = payload.email
+        orguser.user.email = payload.email
     if payload.active is not None:
-        user.active = payload.active
-    if payload.org:
-        org = Org.objects.filter(name=payload.org.name).first()
-        user.org = org
-    user.save()
-    logger.info(f"updated user {user.email}")
-    return user
+        orguser.user.is_active = payload.active
+    orguser.user.save()
+    logger.info(f"updated user {orguser.user.email}")
+    return orguser
 
 
-@adminapi.delete("/organizations/", auth=AdminAuthBearer())
+# ====================================================================================================
+@adminapi.delete("/organizations/", auth=AuthBearer())
 def delete_organization(request, payload: OrgSchema):
-    """Delete an organization"""
+    """delete an organization and all associated org users"""
+    if request.adminuser is None:
+        raise HttpError(400, "unauthorized")
     if request.headers.get("X-DDP-Confirmation") != "yes":
         raise HttpError(400, "missing x-confirmation header")
     org = Org.objects.filter(name=payload.name).first()
     if org:
         for orguser in OrgUser.objects.filter(org=org):
-            logger.warning(f"deleting {org.name} user {orguser.email}")
+            logger.warning(f"deleting {org.name} user {orguser.user.email}")
             orguser.delete()
         org.delete()
         logger.warning(f"deleting {org.name}")
